+++ src/main/java/commands/games/dnd/DndCmd.java
@@ -1,4 +1,4 @@
-package commands.dnd;
+package commands.games.dnd;
 
 public interface DndCmd {
     String description = "Helpful tools when playing Dungeons & Dragons on discord.";
