+++ src/main/java/games/dnd/EncounterGenerator.java
@@ -1,10 +1,11 @@
-package utility;
+package games.dnd;
 
-import models.dnd.Encounter;
-import models.dnd.Monster;
+import games.dnd.models.Encounter;
+import games.dnd.models.Monster;
 import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import utility.Config;
 
 import java.io.BufferedReader;
 import java.io.IOException;
