--- src/main/java/models/BlackjackStates.java
@@ -1,9 +1,0 @@
-package models;
-
-public enum BlackjackStates {
-    PLAYER_WIN,
-    PLAYER_BLACKJACK,
-    DEALER_WIN,
-    DEALER_BLACKJACK,
-    DRAW
-}
